Replace update with meta store rule
* Updating now takes a move, move counts on relevant levels have been
updated

* Cleanup method is gone as it was unnecessarily complex.

* For each store rule in a game, a corresponding meta store rule is
automatically added

+++ src/main/rules/ShiftRule.java
@@ -1,23 +1,24 @@
 package rules;
 
 import base.Config;
-import base.CalculatorGame;
+
 import base.Helpers;
 
 public abstract class ShiftRule extends Rule {
     public CalculatorGame apply(CalculatorGame game) {
         boolean negative = game.getValue() < 0;
-        int[] digits = Helpers.digits((int) game.getValue());
+        int[] digits = Helpers.digits(game.getValue());
         rotate(digits);
         int newValue = valueOf(digits);
         newValue = negative ? -newValue : newValue;
-        return CalculatorGame.generateGame(
-            newValue,
-            game.getGoal(),
-            game.getMovesLeft() - 1,
-            game.getRules(),
-            game.getPortals()
-        );
+        return CalculatorGame
+            .generateGame(
+                newValue,
+                game.getGoal(),
+                game.getMovesLeft() - 1,
+                game.getRules(),
+                game.getPortals()
+            );
     }
 
     public ShiftRule(boolean left) {
